chore: rating 자료형 float -> double

+++ src/main/java/com/naho/umc9th/domain/review/dto/ReviewResDTO.java
@@ -28,7 +28,7 @@ public class ReviewResDTO {
     @Builder
     public record ReviewPreViewDTO(
             String ownerNickname,
-            Float score,
+            Double score,
             String body,
             LocalDate createdAt
     ){}
